Rename release-manager-utils.ts → release-utils.ts
Drop the component-specific prefix so the hooks/ → components/app/
dependency reads as a neutral shared module, not a back-reference to
a specific component.
